requirements.txt
deepgram-sdk>=3.0,<4.0
pyaudio